+++ signin.html
@@ -44,6 +44,6 @@
       </div>
     </div>
   </section>
-  <script src="/js/sign.js" type="module"></script>
+  <script src="/js/signin.js" type="module"></script>
 </body>
 </html>
