+++ signup.html
@@ -33,7 +33,7 @@
         </div>
         <div class="inputModal">
           <label for="doublePassword">비밀번호 확인</label>
-          <input type="password" placeholder="비밀번호" id="doublePassword" autoComplete="off" alt="비밀번호"/>
+          <input type="password" placeholder="비밀번호" id="passwordConfirm" autoComplete="off" alt="비밀번호"/>
           <button class="eyeBtn" type="button">
             <img src="/image/eye-off.svg" class="eyeImg"/>
           </button>
